Use double scores in search index and other improvements

+++ include/InvertedIndex.hpp
@@ -3,4 +3,4 @@
 #include "document.hpp"
 #include "Tokenizer.hpp"
 
-int addDocument(const Document &doc, std::unordered_map<std::string, std::unordered_map<int, int>> &indexes);
+int addDocument(const Document &doc, std::unordered_map<std::string, std::unordered_map<int, double>> &indexes);
