+++ include/Search.hpp
@@ -4,4 +4,4 @@
 #include "Tokenizer.hpp"
 #include "document.hpp"
 
-void start_search(std::vector<Document>& doc_vector, std::unordered_map<std::string, std::unordered_map<int, int>> &index);
+void start_search(std::vector<Document>& doc_vector, std::unordered_map<std::string, std::unordered_map<int, double>> &index);
