+++ src/InvertedIndex.cpp
@@ -3,7 +3,7 @@
 #include "../include/document.hpp"
 #include "../include/Tokenizer.hpp"
 
-int addDocument(const Document &doc, std::unordered_map<std::string, std::unordered_map<int, int>> &indexes)
+int addDocument(const Document &doc, std::unordered_map<std::string, std::unordered_map<int, double>> &indexes)
 {
     auto tokens = tokenize(doc.content);
